fix(ws): pin tip re-fetch to EvmLeaderUpstream
Drop the newHeads header-cache / tip-source registry overbuild. Tip
ownership already lives on per-upstream pollers (SuggestLatestBlock) and
EvmLeaderUpstream; EnforceHighestBlock now UseUpstream-pins the concrete
tip re-fetch to that leader when its LatestBlock covers TipHW.
