+++ architecture/evm/eth_getBlockByNumber.go
@@ -108,39 +108,6 @@ func networkPostForward_eth_getBlockByNumber(ctx context.Context, network common
 	return enforceNonNullBlock(nq, nr)
 }
 
-// observedLatestHeadProvider is implemented by networks that cache the last
-// WS newHeads header before fan-out (erpc.Network). Used to floor HTTP
-// "latest" responses when upstream re-fetch of the tip would fail open.
-type observedLatestHeadProvider interface {
-	LastObservedLatestHead() (blockNumber int64, payload []byte)
-}
-
-// responseFromObservedLatestHead builds an eth_getBlockByNumber response from
-// the cached WS newHeads header when it matches expectedTip. ok is false when
-// the cache is missing, behind, or the network does not expose a tip cache.
-func responseFromObservedLatestHead(network common.Network, nq *common.NormalizedRequest, expectedTip int64) (*common.NormalizedResponse, bool) {
-	provider, ok := network.(observedLatestHeadProvider)
-	if !ok || expectedTip <= 0 {
-		return nil, false
-	}
-	cachedNumber, payload := provider.LastObservedLatestHead()
-	if cachedNumber != expectedTip || len(payload) == 0 {
-		return nil, false
-	}
-	idBytes, err := common.SonicCfg.Marshal(nq.ID())
-	if err != nil {
-		return nil, false
-	}
-	jrr, err := common.NewJsonRpcResponseFromBytes(idBytes, append([]byte(nil), payload...), nil)
-	if err != nil {
-		return nil, false
-	}
-	resp := common.NewNormalizedResponse().
-		WithRequest(nq).
-		WithJsonRpcResponse(jrr)
-	return resp, true
-}
-
 func enforceHighestBlock(ctx context.Context, network common.Network, nq *common.NormalizedRequest, nr *common.NormalizedResponse, re error) (*common.NormalizedResponse, error) {
 	if re != nil {
 		return nr, re
@@ -215,19 +182,6 @@ func enforceHighestBlock(ctx context.Context, network common.Network, nq *common
 			if len(rqj.Params) > 1 {
 				itx, _ = rqj.Params[1].(bool)
 			}
-
-			// Prefer a cached WS newHeads header when the HTTP tip lags. Re-fetching
-			// the concrete tip often fails when only the WS upstream has seen it yet;
-			// pickHighestBlock would then fail-open to the stale response.
-			if !itx {
-				if cached, ok := responseFromObservedLatestHead(network, nq, highestBlockNumber); ok {
-					if nr != nil {
-						nr.Release()
-					}
-					return cached, nil
-				}
-			}
-
 			request, err := BuildGetBlockByNumberRequest(highestBlockNumber, itx)
 			if err != nil {
 				return nil, err
@@ -239,10 +193,18 @@ func enforceHighestBlock(ctx context.Context, network common.Network, nq *common
 			newReq := common.NewNormalizedRequestFromJsonRpcRequest(request)
 			dr := nq.Directives().Clone()
 			dr.SkipCacheRead = "true"
-			// Exclude the stale responder. Upstreams that already have the tip
-			// (via SuggestLatestBlock → LatestBlock) are preferred by
-			// partitionUpstreamsByLatestBlock on the concrete re-fetch.
-			if respBlockNumber > 0 {
+			// Prefer the upstream whose poller already owns this tip
+			// (EvmLeaderUpstream — typically the WS ingress that called
+			// SuggestLatestBlock). Fall back to excluding the stale
+			// responder when no local poller has caught up yet.
+			if leader := network.EvmLeaderUpstream(ctx); leader != nil {
+				if eu, ok := leader.(common.EvmUpstream); ok {
+					if sp := eu.EvmStatePoller(); sp != nil && !sp.IsObjectNull() && sp.LatestBlock() >= highestBlockNumber {
+						dr.UseUpstream = leader.Id()
+					}
+				}
+			}
+			if dr.UseUpstream == "" && respBlockNumber > 0 {
 				dr.UseUpstream = fmt.Sprintf("!%s", nr.UpstreamId())
 			}
 			newReq.SetDirectives(dr)
@@ -254,19 +216,7 @@ func enforceHighestBlock(ctx context.Context, network common.Network, nq *common
 			nnr, err := network.Forward(ctx, newReq)
 			// This is needed in case highest block number is corrupted somehow and for example
 			// it is requesting a very high non-existent block number.
-			picked, pickErr := pickHighestBlock(ctx, nnr, nr, err)
-			// If re-fetch still lost to the stale tip, try the WS-cached header once more
-			// (TipHW may have advanced mid-flight after the first cache check).
-			if !itx && pickErr == nil && picked != nil {
-				_, pickedNumber, refErr := ExtractBlockReferenceFromResponse(ctx, picked)
-				if refErr == nil && pickedNumber < highestBlockNumber {
-					if cached, ok := responseFromObservedLatestHead(network, nq, highestBlockNumber); ok {
-						picked.Release()
-						return cached, nil
-					}
-				}
-			}
-			return picked, pickErr
+			return pickHighestBlock(ctx, nnr, nr, err)
 		} else {
 			return nr, re
 		}
