+++ erpc/subscription_manager.go
@@ -569,16 +569,14 @@ func (h *networkHandle) FinalityDepth() int64 {
 }
 
 // SuggestLatestBlock routes a per-source block observation to the
-// upstream's state poller, then advances the network-level latest tip
-// and caches the newHeads header payload for HTTP tip enforcement.
+// upstream's state poller, then advances the network-level latest tip.
 // sourceId is the ingress adapter's Name(), which for wsupstream.Adapter
 // is "ws:<upstreamId>".
 //
 // Ordering matters: Indexer.Ingest calls this BEFORE fan-out, so by the
 // time any client sees head N on WS, EvmHighestLatestBlockNumber on this
-// pod is already ≥ N and LastObservedLatestHead can serve that header
-// (see Network.NoteObservedLatestHead).
-func (h *networkHandle) SuggestLatestBlock(sourceId string, blockNumber int64, headPayload []byte) {
+// pod is already ≥ N (see Network.NoteObservedLatestBlock).
+func (h *networkHandle) SuggestLatestBlock(sourceId string, blockNumber int64) {
 	const prefix = "ws:"
 	if !strings.HasPrefix(sourceId, prefix) {
 		return
@@ -594,7 +592,7 @@ func (h *networkHandle) SuggestLatestBlock(sourceId string, blockNumber int64, h
 		}
 		break
 	}
-	h.nw.NoteObservedLatestHead(h.nw.appCtx, blockNumber, headPayload)
+	h.nw.NoteObservedLatestBlock(h.nw.appCtx, blockNumber)
 }
 
 // Interface checks: fail the build if either contract drifts.
