+++ indexer/integration_test.go
@@ -107,7 +107,7 @@ type stubNetwork struct {
 
 func (s *stubNetwork) Id() string           { return s.id }
 func (s *stubNetwork) FinalityDepth() int64 { return s.finality }
-func (s *stubNetwork) SuggestLatestBlock(sourceID string, block int64, _ []byte) {
+func (s *stubNetwork) SuggestLatestBlock(sourceID string, block int64) {
 	s.mu.Lock()
 	if s.suggestions == nil {
 		s.suggestions = make(map[string][]int64)
